requirements.txt
babel
beautifulsoup4
patool
portalocker
pydantic
pyshorteners
requests
yt-dlp
ytmusicapi
py-yt-search
